refactor: 폴더 구조 변경 및 헤더 view, controller 분리
