+++ index.html
@@ -14,10 +14,7 @@
     <div class="background"></div>
     <div class="content-layout flex-column-between">
       <header id="header"></header>
-      <main id="main">
-        <form id="transaction-form"></form>
-        <div id="transaction-list"></div>
-      </main>
+      <main id="main"></main>
     </div>
     <script type="module" src="./src/index.js"></script>
   </body>
